add controller delete restrictions

+++ app/controllers/questions_controller.rb
@@ -49,10 +49,15 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
+    unless @question.user == current_user
+      flash[:danger] = "You can only delete your own question."
+      redirect_to root_path 
+    else
 
-    if @question.destroy
-      flash[:success] = "Question was deleted"
-      redirect_to root_path
+      if @question.destroy
+        flash[:success] = "Question was deleted"
+        redirect_to root_path
+      end
     end
     
   end
